Rename Swagger flavor to OpenAPI
 - Use the preferred community naming.
 - Preserve the use of "swagger" where referring to `@nestjs/swagger` (mixin)
