+++ src/flavors/index.ts
@@ -1,2 +1,2 @@
 export { flavor as basic, BasicFlavor } from './basic';
-export { flavor as swagger, SwaggerFlavor } from './swagger';
+export { flavor as openapi, OpenAPIFlavor } from './openapi';
